group d mixed tasks task 5

+++ C++ Lessons/Group_D/Mixed Problems/task4.cpp
@@ -1,13 +1,13 @@
 #include <bits/stdc++.h>
 using namespace std;
 int main() {
-	int n;
+	long long n;
 	cin>>n;
-	int array[n];
+	long long array[n];
 	for(int i = 0; i<n;i++){
 		cin>>array[i];	
 	}
-	int max_element = array[0];
+	long long max_element = array[0];
 	for(int i = 0; i<n;i++){
 		while(max_element < array[i]){
 			max_element = array[i];
